Fix crash in option parsing.
This is reported in Debian bug #716022

When not enough arguments are provided by the user NULL argument is
dereferenced in option parser.

Require one more argument to be non-null since at least two non-option
arguments are required anyway.

This is a cosmetical change - instead of segfault usage message is printed. The
referenced argument is guaranteed to be null and cause crash without the extra
check.